Move ErrorManagerType to subfolder with other types
