+++ src/types/ErrorManagerType.ts
